requirements.txt
# Core FastAPI and server
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.0
python-multipart==0.0.9
python-dotenv==1.0.0
pandas 

# Document processing
PyPDF2==3.0.1
pdfplumber==0.10.3
PyMuPDF==1.23.26
python-docx==1.1.0

# ML and embeddings
torch==2.2.0
transformers==4.38.0
numpy==1.26.4
sentence-transformers==2.2.2

# Vector database
lancedb==0.5.0
pyarrow==15.0.0

# HTTP client
requests==2.31.0

# CLI tools
click==8.1.7

# MCP Tools dependencies
duckduckgo-search==3.9.6

# Dash frontend
dash==2.16.1
dash-bootstrap-components==1.5.0
dash-mantine-components==0.12.1